backend/requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
gitpython
requests
python-dotenv
PyYAML

# Database libraries
SQLAlchemy
psycopg2-binary

# OpenTelemetry dependencies
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
